fix: JPQL DTO로 변환 로직 수정 및 DTO객체 필드 타입 변경

+++ src/main/java/com/itset/itcenteamproject/domain/house/ContractCntDTO.java
@@ -9,5 +9,5 @@ import lombok.Setter;
 @AllArgsConstructor
 public class ContractCntDTO {
     private Integer dongCode;
-    private Integer cnt;
+    private Long cnt;
 }
